third_party/blink/web_tests/platform/mac/external/wpt/html/webappapis/animation-frames/spurious-frame-callbacks-optimization-expected.txt

This is a testharness.js-based test.
[FAIL] Optimization of requestAnimationFrame callbacks that don't modify the DOM shouldn't break animations
  assert_greater_than: expected a number greater than 500 but got 0
Harness: the test ran to completion.
